# doc/openapi.yaml
openapi: 3.1.0
components:
    schemas:
        ACMEUpdateRequest:
            description: ACMEUpdateRequest schema
            properties:
                subdomain:
                    type: string
                txt:
                    type: string
            required:
                - subdomain
                - txt
            type: object
        ACMEUpdateResponse:
            description: ACMEUpdateResponse schema
            properties:
                txt:
                    type: string
            required:
                - txt
            type: object
...
        HTTPError:
            description: HTTPError schema
            properties:
                detail:
                    description: Human readable error message
                    type: string
                errors:
                    items:
                        $ref: '#/components/schemas/ErrorItem'
                    type:
                        - array
                        - "null"
                instance:
                    type: string
                status:
                    description: HTTP status code
                    example: 403
                    type: integer
                title:
                    description: Short title of the error
                    type: string
                type:
                    description: URL of the error type. Can be used to lookup the error in a documentation
                    type: string
            type: object
        LegoHttpDefaultRequest:
            description: LegoHttpDefaultRequest schema
            properties:
                fqdn:
                    type: string
                value:
                    type: string
            required:
                - fqdn
                - value
            type: object
        LegoHttpDefaultResponse:
            description: LegoHttpDefaultResponse schema
            properties:
                fqdn:
                    type: string
                value:
                    type: string
            required:
                - fqdn
                - value
            type: object
        ZMUpdatePTRRequest:
            description: ZMUpdatePTRRequest schema
            properties:
                addresses:
                    items:
                        type: string
                    type: array
                mode:
                    type: string
                target:
                    type: string
            required:
                - addresses
                - target
            type: object
...
        ZMUpdateRequest:
            description: ZMUpdateRequest schema
            properties:
                fqdn:
                    type: string
                ttl:
                    type: integer
                type:
                    type: string
                values:
                    items:
                        type: string
                    type: array
            required:
                - fqdn
                - type
                - values
            type: object
        ZMUpdateResponse:
            description: ZMUpdateResponse schema
            properties:
                changed:
                    type: boolean
                fqdn:
                    type: string
            required:
                - changed
                - fqdn
            type: object
...
        pdnsServer:
            description: pdnsServer schema
            properties:
                config_url:
                    type: string
                daemon_type:
                    type: string
                id:
                    type: string
                type:
                    type: string
                url:
                    type: string
                version:
                    type: string
                zones_url:
                    type: string
            required:
                - config_url
                - daemon_type
                - id
                - type
                - url
                - version
                - zones_url
            type: object
...
        string:
            description: string schema
            type: string
...
    securitySchemes:
        apiKeyAuth:
            description: password
            in: header
            name: X-Api-Key
            type: apiKey
        apiUserAuth:
            description: username
            in: header
            name: X-Api-User
            type: apiKey
        basicAuth:
            scheme: basic
            type: http
        pdnsApiKeyAuth:
            description: static api key
            in: header
            name: X-API-Key
            type: apiKey
info:
    description: HTTP API for zoneomatic DNS updates and PowerDNS-compatible zone management.
    title: Zoneomatic API
    version: 0.0.1
paths:
    /acme/update:
        post:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func4`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nUpdate ACME challenge TXT record"
            operationId: POST_/acme/update
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ACMEUpdateRequest'
                description: Request body for *server.ACMEUpdateRequest
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ACMEUpdateResponse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ACMEUpdateResponse'
                    description: Record updated
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - apiKeyAuth: []
                  apiUserAuth: []
                - basicAuth: []
            summary: update acme
    /api/v1/servers:
        get:
            description: List the forged PowerDNS-compatible server instance.
            operationId: pdnsListServers
            parameters:
                - in: header
                  name: Accept
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                items:
                                    $ref: '#/components/schemas/pdnsServer'
                                type: array
                        application/xml:
                            schema:
                                items:
                                    $ref: '#/components/schemas/pdnsServer'
                                type: array
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - pdnsApiKeyAuth: []
            summary: pdns list servers
    /api/v1/servers/{server_id}:
        get:
            description: Return the forged PowerDNS-compatible server instance.
            operationId: pdnsGetServer
            parameters:
                - in: header
                  name: Accept
                  schema:
                    type: string
                - in: path
                  name: server_id
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/pdnsServer'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/pdnsServer'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - pdnsApiKeyAuth: []
            summary: pdns get server
...
    /cleanup:
        post:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func6`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nClean up ACME challenge TXT record using LEGO HTTP-REQ"
            operationId: POST_/cleanup
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LegoHttpDefaultRequest'
                description: Request body for *server.LegoHttpDefaultRequest
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LegoHttpDefaultResponse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/LegoHttpDefaultResponse'
                    description: Record updated
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - basicAuth: []
            summary: cleanup acme via lego httpreq
...
    /myip:
        get:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func2`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n\n---\n\nReturn client's connection IP address"
            operationId: GET_/myip
            parameters:
                - in: header
                  name: Accept
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/string'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/string'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            summary: myip
    /nic/update:
        get:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func3`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nUpdate DDNS record"
            operationId: GET_/nic/update
            parameters:
                - description: record domain to update
                  in: query
                  name: hostname
                  required: true
                  schema:
                    type: string
                - description: IP address to set
                  in: query
                  name: myip
                  schema:
                    type: string
                - description: IPv6 address to set
                  in: query
                  name: myipv6
                  schema:
                    type: string
                - description: Not supported, a no-op for compatibility.
                  in: query
                  name: offline
                  schema:
                    type: boolean
                - in: header
                  name: Accept
                  schema:
                    type: string
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/string'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/string'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - basicAuth: []
            summary: update ddns
    /present:
        post:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func5`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nUpdate ACME challenge TXT record using LEGO HTTP-REQ"
            operationId: POST_/present
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/LegoHttpDefaultRequest'
                description: Request body for *server.LegoHttpDefaultRequest
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/LegoHttpDefaultResponse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/LegoHttpDefaultResponse'
                    description: Record updated
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - basicAuth: []
            summary: update acme via lego httpreq
    /zm/update:
        post:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func7`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nReplace any existing DNS record value"
            operationId: POST_/zm/update
            parameters:
                - in: header
                  name: Accept
                  schema:
                    type: string
            requestBody:
                content:
                    '*/*':
                        schema:
                            $ref: '#/components/schemas/ZMUpdateRequest'
                description: Request body for server.ZMUpdateRequest
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ZMUpdateResponse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ZMUpdateResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - basicAuth: []
            summary: update any dns record
    /zm/update-ptr:
        post:
            description: "#### Controller: \n\n`github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.func8`\n\n#### Middlewares:\n\n- `github.com/go-fuego/fuego.defaultLogger.middleware`\n- `github.com/vooon/zoneomatic/internal/server.RegisterEndpoints.NewBasicAuthMiddleware.func9`\n\n---\n\nUpdate PTR records in matching reverse zones for the requested addresses, pointing them to the target host. The mode controls whether existing PTR records are appended, replaced or fully synced."
            operationId: POST_/zm/update-ptr
            parameters:
                - in: header
                  name: Accept
                  schema:
                    type: string
            requestBody:
                content:
                    '*/*':
                        schema:
                            $ref: '#/components/schemas/ZMUpdatePTRRequest'
                description: Request body for server.ZMUpdatePTRRequest
                required: true
            responses:
                "200":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ZMUpdatePTRResponse'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/ZMUpdatePTRResponse'
                    description: OK
                "400":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Bad Request _(validation or deserialization error)_
                "500":
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                        application/xml:
                            schema:
                                $ref: '#/components/schemas/HTTPError'
                    description: Internal Server Error _(panics)_
            security:
                - basicAuth: []
            summary: update ptr records